requirements.txt
cffi==1.15.1
charset-normalizer==3.1.0
colorthief==0.2.1
cryptography==40.0.2
darkdetect==0.8.0
lxml==4.9.2
numpy==1.24.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.0.3
PyQt-Fluent-Widgets==0.8.8
PyQt5-Frameless-Window==0.2.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQt5-stubs==5.15.6.0
PyQt5==5.15.9
python-docx==0.8.11
python-pptx==0.6.21
pywin32==306
scipy==1.10.1
setuptools==60.2.0
Wand==0.6.11
wheel==0.37.1
XlsxWriter==3.1.2
